split out parameter functions
