+++ attach.c
@@ -46,6 +46,7 @@
 #include "ncrypt/ncrypt.h"
 #include "options.h"
 #include "pager.h"
+#include "parameter.h"
 #include "protos.h"
 #include "rfc1524.h"
 #include "state.h"
