+++ handler.c
@@ -50,6 +50,7 @@
 #include "mutt_curses.h"
 #include "ncrypt/ncrypt.h"
 #include "options.h"
+#include "parameter.h"
 #include "protos.h"
 #include "rfc1524.h"
 #include "rfc3676.h"
