+++ ncrypt/crypt.c
@@ -49,6 +49,7 @@
 #include "mutt_curses.h"
 #include "ncrypt.h"
 #include "options.h"
+#include "parameter.h"
 #include "protos.h"
 #include "rfc822.h"
 #include "state.h"
