+++ ncrypt/smime.c
@@ -52,6 +52,7 @@
 #include "mutt_menu.h"
 #include "ncrypt.h"
 #include "options.h"
+#include "parameter.h"
 #include "protos.h"
 #include "rfc822.h"
 #include "state.h"
