+++ parameter.h
@@ -40,4 +40,9 @@ static inline struct Parameter *mutt_new_parameter(void)
   return safe_calloc(1, sizeof(struct Parameter));
 }
 
+void mutt_delete_parameter(const char *attribute, struct Parameter **p);
+void mutt_set_parameter(const char *attribute, const char *value, struct Parameter **p);
+void mutt_free_parameter(struct Parameter **p);
+char *mutt_get_parameter(const char *s, struct Parameter *p);
+
 #endif /* _MUTT_PARAMETER_H */
