+++ rfc3676.c
@@ -39,6 +39,7 @@
 #include "lib.h"
 #include "mutt_curses.h"
 #include "options.h"
+#include "parameter.h"
 #include "protos.h"
 #include "state.h"
