+++ send.c
@@ -52,6 +52,7 @@
 #include "mutt_regex.h"
 #include "ncrypt/ncrypt.h"
 #include "options.h"
+#include "parameter.h"
 #include "pattern.h"
 #include "protos.h"
 #include "rfc2047.h"
